backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-telegram-bot
upstox-python-sdk
requests
numpy
pandas
nsetools
PyJWT
